Team Register, 다대다 관계를 집어넣고 Result에서 성공, 실패 두가지 결과가 나오는 오류 발생

+++ src/main/java/com/example/rascalserver/DAO/AccountJpaRepo.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AccountJpaRepo extends JpaRepository<Account, Integer> {
+public interface AccountJpaRepo extends JpaRepository<Account, Long> {
     Optional<Account> findByEmail(String email);
     Optional<Account> findByUid(Long uid);
 }
