+++ src/main/java/com/example/rascalserver/DAO/TeamJpaRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeamJpaRepo extends JpaRepository<Team, Integer> {
+public interface TeamJpaRepo extends JpaRepository<Team, Long> {
 }
